+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - GTRANSFORM-213 - add labels to textareas identified in accessibility audit
 - GTRANSFORM-238 - page heading fixes
 - GTRANSFORM-238 - use page heading in the page title
-- XOT-914 - Add search sort order healthcheck
+- XOT-914 - Add search sort order healthcheck with feature flag
 - XOT-1034 - Add new landing page, remove obsolete international page styling, update metadata for cms pages
 - TT-1808: Update directory components to add "no-validate" no cache middleware
 - No ticket - Add trade barriers to search